Add fractals
Added several classes and scripts to add the fractals. There are still mismatches between the wanted(and expected) result and the actual result.
It loads up the configuration in the editor at the beginning of the scene. But, it is possible to refresh(and regenerate) the fractals at runtime by pressing first the right button of the mouse and then the left one(after changing the vales).